extract plugin settings, add edition property
